OX-HUGO: cventry now at org heading level
Previously a cventry was inside a div and forced a h3 heading. The content
would not me inside the div because on markdown it wouldn't be
rendered. For limitation it is better to drop the div, make the heading
respect the org level and the properties list is just in one UL tag